+++ lib/npc/ally/simple_ally.dart
@@ -1,23 +1,39 @@
-import 'package:bonfire/enemy/enemy.dart';
+import 'package:bonfire/npc/ally/ally.dart';
 import 'package:bonfire/util/direction.dart';
 import 'package:bonfire/util/direction_animations/simple_direction_animation.dart';
+import 'package:bonfire/util/mixins/attackable.dart';
 import 'package:bonfire/util/mixins/direction_animation.dart';
 import 'package:flame/components.dart';
 
+///
+/// Created by
+///
+/// ─▄▀─▄▀
+/// ──▀──▀
+/// █▀▀▀▀▀█▄
+/// █░░░░░█─█
+/// ▀▄▄▄▄▄▀▀
+///
+/// Rafaelbarbosatec
+/// on 24/03/22
+
 /// Enemy with animation in all direction
-class SimpleEnemy extends Enemy with DirectionAnimation {
-  SimpleEnemy({
+class SimpleAlly extends Ally with DirectionAnimation {
+  SimpleAlly({
     required Vector2 position,
     required Vector2 size,
     required SimpleDirectionAnimation animation,
     double life = 100,
     double speed = 100,
     Direction initDirection = Direction.right,
+    ReceivesAttackFromEnum receivesAttackFrom =
+        ReceivesAttackFromEnum.PLAYER_AND_ALLY,
   }) : super(
           position: position,
           size: size,
           life: life,
           speed: speed,
+          receivesAttackFrom: receivesAttackFrom,
         ) {
     this.animation = animation;
     lastDirection = initDirection;