+++ lib/npc/enemy/rotation_enemy.dart
@@ -14,11 +14,14 @@ class RotationEnemy extends Enemy with UseSpriteAnimation, UseAssetsLoader {
     double currentRadAngle = -1.55,
     double speed = 100,
     double life = 100,
+    ReceivesAttackFromEnum receivesAttackFrom =
+        ReceivesAttackFromEnum.PLAYER_AND_ALLY,
   }) : super(
           position: position,
           size: size,
           life: life,
           speed: speed,
+          receivesAttackFrom: receivesAttackFrom,
         ) {
     angle = currentRadAngle;
     loader?.add(AssetToLoad(animIdle, (value) {