+++ lib/util/extensions/enemy/enemy_extensions.dart
@@ -1,70 +1,18 @@
-import 'dart:math';
-
 import 'package:bonfire/collision/collision_config.dart';
-import 'package:bonfire/collision/object_collision.dart';
-import 'package:bonfire/enemy/enemy.dart';
 import 'package:bonfire/lighting/lighting_config.dart';
+import 'package:bonfire/npc/enemy/enemy.dart';
 import 'package:bonfire/player/player.dart';
 import 'package:bonfire/util/direction.dart';
 import 'package:bonfire/util/extensions/game_component_extensions.dart';
 import 'package:bonfire/util/extensions/movement_extensions.dart';
+import 'package:bonfire/util/extensions/npc/npc_extensions.dart';
 import 'package:flame/components.dart';
 import 'package:flutter/widgets.dart';
 
+import '../../mixins/attackable.dart';
+
 /// Functions util to use in your [Enemy]
 extension EnemyExtensions on Enemy {
-  /// This method we notify when detect the player when enter in [radiusVision] configuration
-  /// Method that bo used in [update] method.
-  void seePlayer({
-    required Function(Player) observed,
-    VoidCallback? notObserved,
-    double radiusVision = 32,
-  }) {
-    Player? player = gameRef.player;
-    if (player == null || player.isDead) {
-      notObserved?.call();
-      return;
-    }
-    this.seeComponent(
-      player,
-      observed: (c) => observed(c as Player),
-      notObserved: notObserved,
-      radiusVision: radiusVision,
-    );
-  }
-
-  /// Checks whether the player is within range. If so, move to it.
-  void seeAndMoveToPlayer({
-    required Function(Player) closePlayer,
-    VoidCallback? notObserved,
-    VoidCallback? observed,
-    double radiusVision = 32,
-    double margin = 10,
-    bool runOnlyVisibleInScreen = true,
-  }) {
-    if (isDead) return;
-    if (runOnlyVisibleInScreen && !this.isVisible) return;
-
-    seePlayer(
-      radiusVision: radiusVision,
-      observed: (player) {
-        observed?.call();
-        this.followComponent(
-          player,
-          dtUpdate,
-          closeComponent: (comp) => closePlayer(comp as Player),
-          margin: margin,
-        );
-      },
-      notObserved: () {
-        if (!this.isIdle) {
-          this.idle();
-        }
-        notObserved?.call();
-      },
-    );
-  }
-
   ///Execute simple attack melee using animation
   void simpleAttackMelee({
     required double damage,
@@ -97,6 +45,7 @@ extension EnemyExtensions on Enemy {
       animationDown: animationDown,
       animationLeft: animationLeft,
       animationRight: animationRight,
+      attackFrom: AttackFromEnum.ENEMY,
     );
 
     execute?.call();
@@ -146,6 +95,7 @@ extension EnemyExtensions on Enemy {
       destroySize: destroySize,
       lightingConfig: lightingConfig,
       enableDiagonal: enableDiagonal,
+      attackFrom: AttackFromEnum.ENEMY,
     );
 
     if (execute != null) execute();
@@ -190,42 +140,5 @@ extension EnemyExtensions on Enemy {
         notObserved?.call();
       },
     );
-  }
-
-  /// Get angle between enemy and player
-  /// player as a base
-  double getAngleFomPlayer() {
-    Player? player = this.gameRef.player;
-    if (player == null) return 0.0;
-    return atan2(
-      playerRect.center.dy - enemyRect.center.dy,
-      playerRect.center.dx - enemyRect.center.dx,
-    );
-  }
-
-  /// Get angle between enemy and player
-  /// enemy position as a base
-  double getInverseAngleFomPlayer() {
-    Player? player = this.gameRef.player;
-    if (player == null) return 0.0;
-    return atan2(
-      this.position.y - playerRect.center.dy,
-      this.position.x - playerRect.center.dx,
-    );
-  }
-
-  /// Gets player position used how base in calculations
-  Rect get playerRect {
-    return (gameRef.player is ObjectCollision
-            ? (gameRef.player as ObjectCollision).rectCollision
-            : gameRef.player?.toRect()) ??
-        Rect.zero;
-  }
-
-  /// Gets enemy position used how base in calculations
-  Rect get enemyRect {
-    return (this.isObjectCollision()
-        ? (this as ObjectCollision).rectCollision
-        : toRect());
   }
 }