+++ lib/widgets/bonfire_tiled_widget.dart
@@ -44,6 +44,18 @@ class BonfireTiledWidget extends StatefulWidget {
   /// Represents a map (or world) where the game occurs.
   final TiledWorldMap map;
 
+  /// The [FocusNode] to control the games focus to receive event inputs.
+  /// If omitted, defaults to an internally controlled focus node.
+  final FocusNode? focusNode;
+
+  /// Whether the [focusNode] requests focus once the game is mounted.
+  /// Defaults to true.
+  final bool autofocus;
+
+  /// Initial mouse cursor for this [GameWidget]
+  /// mouse cursor can be changed in runtime using [Game.mouseCursor]
+  final MouseCursor? mouseCursor;
+
   final TapInGame? onTapDown;
   final TapInGame? onTapUp;
 
@@ -54,7 +66,7 @@ class BonfireTiledWidget extends StatefulWidget {
   final Widget? progress;
   final CameraConfig? cameraConfig;
   final AnimatedSwitcherTransitionBuilder transitionBuilder;
-  final Duration? progressTransitionDuration;
+  final Duration progressTransitionDuration;
   final GameColorFilter? colorFilter;
 
   const BonfireTiledWidget({
@@ -74,7 +86,7 @@ class BonfireTiledWidget extends StatefulWidget {
     this.progress,
     this.cameraConfig,
     this.transitionBuilder = AnimatedSwitcher.defaultTransitionBuilder,
-    this.progressTransitionDuration,
+    this.progressTransitionDuration = const Duration(milliseconds: 500),
     this.colorFilter,
     this.components,
     this.overlayBuilderMap,
@@ -82,6 +94,9 @@ class BonfireTiledWidget extends StatefulWidget {
     this.onTapDown,
     this.onTapUp,
     this.onReady,
+    this.focusNode,
+    this.autofocus = true,
+    this.mouseCursor,
   }) : super(key: key);
   @override
   _BonfireTiledWidgetState createState() => _BonfireTiledWidgetState();
@@ -122,17 +137,21 @@ class _BonfireTiledWidgetState extends State<BonfireTiledWidget>
   Widget build(BuildContext context) {
     return Stack(
       children: [
-        _buildGame(),
+        if (_game != null)
+          CustomGameWidget(
+            game: _game!,
+            overlayBuilderMap: widget.overlayBuilderMap,
+            initialActiveOverlays: widget.initialActiveOverlays,
+            mouseCursor: widget.mouseCursor,
+            autofocus: widget.autofocus,
+            focusNode: widget.focusNode,
+          ),
         StreamBuilder<bool>(
           stream: _loadingStream.stream,
           builder: (context, snapshot) {
-            bool _loading = true;
-            if (snapshot.hasData) {
-              _loading = snapshot.data!;
-            }
+            bool _loading = !snapshot.hasData || snapshot.data == true;
             return AnimatedSwitcher(
-              duration: widget.progressTransitionDuration ??
-                  Duration(milliseconds: 500),
+              duration: widget.progressTransitionDuration,
               transitionBuilder: widget.transitionBuilder,
               child: _loading ? _defaultProgress() : SizedBox.shrink(),
             );
@@ -146,35 +165,36 @@ class _BonfireTiledWidgetState extends State<BonfireTiledWidget>
     try {
       TiledWorldData tiled = await widget.map.build();
 
-      List<GameComponent> components = (tiled.components ?? []);
-      if (widget.components != null) components.addAll(widget.components!);
-      _game = BonfireGame(
-        context: context,
-        joystickController: widget.joystick,
-        player: widget.player,
-        interface: widget.interface,
-        map: tiled.map,
-        components: components,
-        background: widget.background,
-        constructionMode: widget.constructionMode,
-        showCollisionArea: widget.showCollisionArea,
-        showFPS: widget.showFPS,
-        gameController: widget.gameController,
-        constructionModeColor:
-            widget.constructionModeColor ?? Colors.cyan.withOpacity(0.5),
-        collisionAreaColor: widget.collisionAreaColor ??
-            Colors.lightGreenAccent.withOpacity(0.5),
-        lightingColorGame: widget.lightingColorGame,
-        cameraConfig: widget.cameraConfig,
-        colorFilter: widget.colorFilter,
-        onTapDown: widget.onTapDown,
-        onTapUp: widget.onTapUp,
-        onReady: (game) {
-          _showProgress(false);
-          widget.onReady?.call(game);
-        },
-      );
-      setState(() {});
+      setState(() {
+        List<GameComponent> components = (tiled.components ?? []);
+        components.addAll(widget.components ?? []);
+        _game = BonfireGame(
+          context: context,
+          joystickController: widget.joystick,
+          player: widget.player,
+          interface: widget.interface,
+          map: tiled.map,
+          components: components,
+          background: widget.background,
+          constructionMode: widget.constructionMode,
+          showCollisionArea: widget.showCollisionArea,
+          showFPS: widget.showFPS,
+          gameController: widget.gameController,
+          constructionModeColor:
+              widget.constructionModeColor ?? Colors.cyan.withOpacity(0.5),
+          collisionAreaColor: widget.collisionAreaColor ??
+              Colors.lightGreenAccent.withOpacity(0.5),
+          lightingColorGame: widget.lightingColorGame,
+          cameraConfig: widget.cameraConfig,
+          colorFilter: widget.colorFilter,
+          onTapDown: widget.onTapDown,
+          onTapUp: widget.onTapUp,
+          onReady: (game) {
+            _showProgress(false);
+            widget.onReady?.call(game);
+          },
+        );
+      });
     } catch (e) {
       print('(BonfireTiledWidget) Error: $e');
     }
@@ -188,15 +208,6 @@ class _BonfireTiledWidgetState extends State<BonfireTiledWidget>
             child: CircularProgressIndicator(),
           ),
         );
-  }
-
-  Widget _buildGame() {
-    if (_game == null) return SizedBox.shrink();
-    return CustomGameWidget(
-      game: _game!,
-      overlayBuilderMap: widget.overlayBuilderMap,
-      initialActiveOverlays: widget.initialActiveOverlays,
-    );
   }
 
   void _showProgress(bool show) async {